Mobile_Subscriber_BE/requirements.txt
asgiref==3.2.10
autopep8==1.6.0
bleach==4.1.0
Django==3.1.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.12.1
djangorestframework-simplejwt==5.1.0
drf-api-logger==1.0.10
packaging==21.3
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
webencodings==0.5.1
